First stab at making boundary conditions truly free
We apply partitioning to the innermost loop only. This is a terrible
decision when the innermost loop is a small RDom, e.g. iterating over a
convolution kernel. This changes it to apply to the outermost loop it
can, and to operate recursively on the steady state.

So a 2D boundary condition now produces five cases instead of three. The
cases are the top, left, right, bottom, and middle of the image. The
corners of the image get included in the top and bottom or left and
right depending on what the nesting order of x and y were.