Fix bug in partition loops
The steady state exprs were depending on the loop var itself, which
means the solve didn't successfully move the loop var to the LHS! It can
appear on the RHS once the bounds are taken over the scope (the scope
might include inner loop vars for which the bounds depend on the loop
var being split).

This was previously supposedly tracked in free_vars, but this was a dead
member (never read from).